stop intercepting INT signals (ctrl-C)
this signal hook provoked several problems in interactive applications
as the devuan -sdk where it may be desirable and almost automatic
behavior to stop an operation with ctrl-c. intercepting this signal in
endgame then leads to frustration as all variables and current
configurations are forgotten.

it must be still assessed carefully the impact of this change in all
applications and eventually reverted to have an "interactive" mode
where zuper avoids interception.